First checkpoint for adding quickcheck tests. (Warning, tree kind of broken, see below)
Things are TBH, pretty broken right now and there's a lot of extra debug code in place. Also I think building through Xcode might be broken.

Regardless, we have a good first stab at getting QuickCheck to do roundtripping of complex types of data, and writing instances of Arbitrary for other types should be pretty easy as well.

The main problem right now is that there a number of crashes when running under QuickCheck, which I'm working to resolve. This commit is basically just a checkpoint.